Model directory with API models and (eventually) interfaces.
Not exactly sure how this spread is going to work. I know this is
the set of data classes which will be used to interact with the
game from all implementations. I don't know if there is an
internal model as well and if so where the conversions happen. I
don't know what happens with the fact that presumably I want to
let people use the innards to see updated state on their own (for
example to test out moves) and what that means since that service
would not use these external API models... I suspect it will sort
itself out if I let it sit for a while!

Additionally I'm getting real anxious on this gem map thing. I
like the idea of having very easily serializable types on all the
API models. I also like the idea of having set resource types for
a schema, which would likely make things tighter than the uber
flexible gem map. This one I am really not sure about.